Update the code to use import in replicated indicators.
