+++ agents_indicators/views.py
@@ -6,6 +6,9 @@ from .models import PercentAgentsPerAreaOperation
 from datetime import datetime
 import json
 from celery.decorators import task
+from quero_cultura.views import build_temporal_indicator
+from quero_cultura.views import build_operation_area_indicator
+
 
 def index(request):
 
@@ -73,36 +76,6 @@ def index(request):
     return render(request, 'agents_indicators/agents_indicators.html', context)
 
 
-def build_temporal_indicator(new_data, old_data):
-    temporal_indicator = {}
-
-    for agent in new_data:
-        split_date = agent["createTimestamp"]["date"].split("-")
-
-        year = split_date[0]
-        month = split_date[1]
-
-        if not (year in temporal_indicator):
-            temporal_indicator[year] = {}
-            temporal_indicator[year][month] = 1
-        elif not (month in temporal_indicator.get(year)):
-            temporal_indicator[year][month] = 1
-        else:
-            temporal_indicator[year][month] += 1
-
-    for year in old_data:
-        if not (year in temporal_indicator):
-            temporal_indicator[year] = old_data[year]
-        else:
-            for month in old_data[year]:
-                if not (month in temporal_indicator[year]):
-                    temporal_indicator[year][month] = old_data[year][month]
-                else:
-                    temporal_indicator[year][month] += old_data[year][month]
-
-    return temporal_indicator
-
-
 def build_type_indicator(new_data):
     per_type = {}
 
@@ -113,25 +86,6 @@ def build_type_indicator(new_data):
             per_type[agent["type"]["name"]] += 1
 
     return per_type
-
-
-def build_operation_area_indicator(new_data, old_data):
-    per_operation_area = {}
-
-    for agent in new_data:
-        for area in agent["terms"]["area"]:
-            if not (area in per_operation_area):
-                per_operation_area[area] = 1
-            else:
-                per_operation_area[area] += 1
-
-    for area in old_data:
-            if not (area in per_operation_area):
-                per_operation_area[area] = old_data[area]
-            else:
-                per_operation_area[area] += old_data[area]
-
-    return per_operation_area
 
 
 @task(name="update_agent_indicator")
